requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
python-dotenv>=1.0.0
requests>=2.31.0
werkzeug>=3.0.0

# Log monitoring
watchdog>=3.0.0

# PDF report generation
reportlab>=4.0.0

# Charting (optional - for matplotlib charts)
matplotlib>=3.8.0
seaborn>=0.13.0

# Machine learning — stay on NumPy 1.x (match Thonny / your interpreter); keep sklearn aligned
numpy>=1.26.0,<2
scikit-learn>=1.3.0,<1.6
pandas>=2.0.0

# Association rule mining (Apriori) — SOC dashboard / teacher demo
mlxtend>=0.22.0

# AI explanation (optional)
openai>=1.10.0

# GeoIP location-based risk (optional)
geoip2>=4.7.0

# PDF/data processing
openpyxl>=3.1.0

# Technical report (Word) — run: py generate_report.py
python-docx>=1.1.0
Pillow>=10.0.0

# Real dashboard screenshots in report (optional): after install, run: playwright install chromium
playwright>=1.40.0

# Viva deck: py generate_viva_presentation.py
python-pptx>=0.6.21
